12_2 (логирование main.py)
